requirements.txt
streamlit
pillow
matplotlib
pandas
numpy